Replace usage of raw std::cin with CommandReader
Reading data directly from std::cin confused CommandReader. For example
reading an std::string when it was intended to get a line of input from
the user did not consume the new line character, thus the next call to
CommandReader::new_line read the rest of the previous line instead of
the next one.